+++ openeo_driver/dummy/dummy_backend.py
@@ -26,6 +26,7 @@ from openeo.metadata import (
     CollectionMetadata,
     SpatialDimension,
     TemporalDimension,
+    CubeMetadata,
 )
 from openeo.util import rfc3339
 from openeo.utils.version import ComparableVersion
@@ -202,7 +203,7 @@ def mock_side_effect(fun):
 
 class DummyDataCube(DriverDataCube):
 
-    def __init__(self, metadata: CollectionMetadata = None):
+    def __init__(self, metadata: Optional[CubeMetadata] = None):
         super(DummyDataCube, self).__init__(metadata=metadata)
 
         # TODO #47: remove this non-standard process?
@@ -227,15 +228,15 @@ class DummyDataCube(DriverDataCube):
     def reduce_dimension(
         self, reducer, *, dimension: str, context: Optional[dict] = None, env: EvalEnv
     ) -> "DummyDataCube":
-        return DummyDataCube(self.metadata.reduce_dimension(dimension_name=dimension))
+        return DummyDataCube(metadata=self.metadata.reduce_dimension(dimension_name=dimension))
 
     @mock_side_effect
     def add_dimension(self, name: str, label, type: str = "other") -> 'DummyDataCube':
-        return DummyDataCube(self.metadata.add_dimension(name=name, label=label, type=type))
+        return DummyDataCube(metadata=self.metadata.add_dimension(name=name, label=label, type=type))
 
     @mock_side_effect
     def drop_dimension(self, name: str) -> 'DriverDataCube':
-        return DummyDataCube(self.metadata.drop_dimension(name=name))
+        return DummyDataCube(metadata=self.metadata.drop_dimension(name=name))
 
     @mock_side_effect
     def dimension_labels(self, dimension: str) -> 'DriverDataCube':
@@ -451,10 +452,9 @@ class DummyCatalog(CollectionCatalog):
                 "t": {"type": "temporal", "extent": ["2019-01-01", None]},
                 "bands": {"type": "bands", "values": ["MAP"]},
             },
-            'summaries': {
-                "eo:bands": [
-                    {"name": "MAP", "common_name": "blue"}
-                ]
+            "summaries": {
+                "eo:bands": [{"name": "MAP", "common_name": "blue"}],
+                "bands": [{"name": "MAP", "eo:common_name": "blue"}],
             },
             'links': [],
             '_private': {'password': 'dragon'},
@@ -481,11 +481,15 @@ class DummyCatalog(CollectionCatalog):
                 "t": {"type": "temporal", "extent": ["2019-01-01", None]},
                 "bands": {"type": "bands", "values": ["VV","VH"]},
             },
-            'summaries': {
+            "summaries": {
                 "eo:bands": [
                     {"name": "VV", "common_name": "blue"},
-                    {"name": "VH", "common_name": "blue"}
-                ]
+                    {"name": "VH", "common_name": "blue"},
+                ],
+                "bands": [
+                    {"name": "VV", "eo:common_name": "blue"},
+                    {"name": "VH", "eo:common_name": "blue"},
+                ],
             },
             'links': [],
             '_private': {'password': 'dragon'}
@@ -521,6 +525,16 @@ class DummyCatalog(CollectionCatalog):
                     {"name": "SCL", "common_name": "nir"},
                     {"name": "sunAzimuthAngles", "common_name": "nir"},
                     {"name": "sunZenithAngles", "common_name": "nir"},
+                ],
+                "bands": [
+                    {"name": "B02", "eo:common_name": "blue"},
+                    {"name": "B03", "eo:common_name": "green"},
+                    {"name": "B04", "eo:common_name": "red"},
+                    {"name": "B08", "eo:common_name": "nir"},
+                    {"name": "CLP", "eo:common_name": "nir"},
+                    {"name": "SCL", "eo:common_name": "nir"},
+                    {"name": "sunAzimuthAngles", "eo:common_name": "nir"},
+                    {"name": "sunZenithAngles", "eo:common_name": "nir"},
                 ]
             },
             'links': [],
@@ -553,6 +567,12 @@ class DummyCatalog(CollectionCatalog):
                     {"name": "B03", "common_name": "green"},
                     {"name": "B04", "common_name": "red"},
                     {"name": "B08", "common_name": "nir"},
+                ],
+                "bands": [
+                    {"name": "B02", "eo:common_name": "blue"},
+                    {"name": "B03", "eo:common_name": "green"},
+                    {"name": "B04", "eo:common_name": "red"},
+                    {"name": "B08", "eo:common_name": "nir"},
                 ]
             },
             'links': [],
@@ -586,8 +606,12 @@ class DummyCatalog(CollectionCatalog):
             "summaries": {
                 "eo:bands": [
                     {"name": "FAPAR_10M"},
-                    {"name": "SCENECLASSIFICATION_20M"}
-                ]
+                    {"name": "SCENECLASSIFICATION_20M"},
+                ],
+                "bands": [
+                    {"name": "FAPAR_10M"},
+                    {"name": "SCENECLASSIFICATION_20M"},
+                ],
             },
             "_vito": {
                 "properties": {
@@ -802,7 +826,7 @@ class DummyBatchJobs(BatchJobs):
                     "roles": ["data"],
                     "type": "text/csv",
                     "href": "s3://OpenEO-data/batch_jobs/j-2406047c20fc4966ab637d387502728f/timeseries.csv",
-                    "bands": [Band(name="S2-L2A-EVI_t0"), Band(name="S2-L2A-EVI_t1"), Band(name="S2-L2A-EVI_t2")],
+                    "bands": [Band(name="S2-L2A-EVI_t0"), Band(name="S2-L2A-EVI_t1"), {"name": "S2-L2A-EVI_t2"}],
                     "geometry": {"type": "Polygon",
                                  "coordinates": [[[2.70964374625748, 51.00377983772219],
                                                   [2.70964374625748, 51.10589339112414],
@@ -855,7 +879,7 @@ class DummyBatchJobs(BatchJobs):
                     "href": f"{self._output_root()}/{job_id}/subfolder/output.tiff",
                     "type": "image/tiff; application=geotiff",
                     "roles": ["data"],
-                    "bands": [Band(name="NDVI", common_name="NDVI", wavelength_um=1.23)],
+                    "bands": [Band(name="NDVI", wavelength_um=1.23)],
                     "nodata": 123,
                     "instruments": "MSI",
                     "bbox": [0.0, 50.0, 5.0, 55.0],
@@ -872,7 +896,7 @@ class DummyBatchJobs(BatchJobs):
                 "href": f"{self._output_root()}/{job_id}/output.tiff",
                 "type": "image/tiff; application=geotiff",
                 "roles": ["data"],
-                "bands": [Band(name="NDVI", common_name="NDVI", wavelength_um=1.23)],
+                "bands": [Band(name="B02", common_name="blue", wavelength_um=0.665)],
                 "nodata": 123,
                 "instruments": "MSI",
                 "bbox": [0.0, 50.0, 5.0, 55.0],
@@ -1068,8 +1092,7 @@ class DummyBackendImplementation(OpenEoBackendImplementation):
             self, format: str, glob_pattern: str, options: dict, load_params: LoadParameters, env: EvalEnv
     ) -> DummyDataCube:
         _register_load_collection_call(glob_pattern, load_params)
-        metadata = CollectionMetadata(
-            {},
+        metadata = CubeMetadata(
             dimensions=[
                 SpatialDimension(name="x", extent=[]),
                 SpatialDimension(name="y", extent=[]),